feat(PE-3065) transfer ant config
