+++ src/components/layout/TransactionWorkflow/TransactionWorkflow.tsx
@@ -274,6 +274,7 @@ function TransactionWorkflow({
             case ANT_INTERACTION_TYPES.SET_TARGET_ID:
             case ANT_INTERACTION_TYPES.SET_TICKER:
             case ANT_INTERACTION_TYPES.SET_CONTROLLER:
+            case ANT_INTERACTION_TYPES.TRANSFER:
             case ANT_INTERACTION_TYPES.SET_NAME: {
               return {
                 pending: {
