+++ src/components/modals/ManageAntModal/ManageAntModal.tsx
@@ -25,27 +25,7 @@ import { mapKeyToAttribute } from '../../cards/AntCard/AntCard';
 import { ArrowLeft, CloseIcon, PencilIcon } from '../../icons';
 import ValidationInput from '../../inputs/text/ValidationInput/ValidationInput';
 import TransactionStatus from '../../layout/TransactionStatus/TransactionStatus';
-
-const EDITABLE_FIELDS = [
-  'name',
-  'ticker',
-  'targetID',
-  'ttlSeconds',
-  'controller',
-];
-
-const ACTION_FIELDS: {
-  [x: string]: {
-    title: string;
-    fn?: () => void;
-  };
-} = {
-  owner: {
-    title: TRANSACTION_TYPES.TRANSFER,
-    // todo: navigate to /transfer route
-    fn: () => alert('coming soon!'),
-  },
-};
+import TransferAntModal from '../TransferAntModal/TransferAntModal';
 
 function ManageAntModal() {
   const { id } = useParams();
@@ -58,6 +38,29 @@ function ManageAntModal() {
   const [editingField, setEditingField] = useState<string>();
   const [modifiedValue, setModifiedValue] = useState<string | number>();
   const [rows, setRows] = useState<ManageAntRow[]>([]);
+  const [showTransferAntModal, setShowTransferAntModal] =
+    useState<boolean>(false);
+
+  const EDITABLE_FIELDS = [
+    'name',
+    'ticker',
+    'targetID',
+    'ttlSeconds',
+    'controller',
+  ];
+
+  const ACTION_FIELDS: {
+    [x: string]: {
+      title: string;
+      fn?: () => void;
+    };
+  } = {
+    owner: {
+      title: TRANSACTION_TYPES.TRANSFER,
+      // todo: navigate to /transfer route
+      fn: () => setShowTransferAntModal(true),
+    },
+  };
 
   useEffect(() => {
     try {
@@ -129,236 +132,255 @@ function ManageAntModal() {
   }
 
   return (
-    <div className="page">
-      <div className="flex-row flex-space-between">
-        <span className="flex white text-large bold">
+    <>
+      <div className="page">
+        <div className="flex-row flex-space-between">
+          <span className="flex white text-large bold">
+            <button
+              className="faded text-large bold underline link center"
+              onClick={() => navigate('/manage/ants')}
+            >
+              <ArrowLeft
+                width={30}
+                height={20}
+                viewBox={'0 0 20 20'}
+                fill={'var(--text-white)'}
+              />
+              Manage ANTs
+            </button>
+            <Tooltip
+              placement="right"
+              title={id}
+              showArrow={true}
+              overlayStyle={{
+                maxWidth: 'fit-content',
+              }}
+            >
+              <span>&nbsp;/&nbsp;{antName}</span>
+            </Tooltip>
+          </span>
+          {/* TODO: make sure the table doesn't refresh if no actions were saved/written */}
           <button
-            className="faded text-large bold underline link center"
+            className="flex flex-right pointer"
             onClick={() => navigate('/manage/ants')}
           >
-            <ArrowLeft
-              width={30}
-              height={20}
-              viewBox={'0 0 20 20'}
-              fill={'var(--text-white)'}
-            />
-            Manage ANTs
+            <CloseIcon width="30px" height={'30px'} fill="var(--text-white)" />
           </button>
-          <Tooltip
-            placement="right"
-            title={id}
-            showArrow={true}
-            overlayStyle={{
-              maxWidth: 'fit-content',
-            }}
-          >
-            <span>&nbsp;/&nbsp;{antName}</span>
-          </Tooltip>
-        </span>
-        {/* TODO: make sure the table doesn't refresh if no actions were saved/written */}
-        <button
-          className="flex flex-right pointer"
-          onClick={() => navigate('/manage/ants')}
-        >
-          <CloseIcon width="30px" height={'30px'} fill="var(--text-white)" />
-        </button>
-      </div>
-      <div className="flex-row">
-        <Table
-          showHeader={false}
-          style={{ width: '100%' }}
-          onRow={(row: ManageAntRow) => ({
-            className: row.attribute === editingField ? 'active-row' : '',
-          })}
-          scroll={{ x: true }}
-          columns={[
-            {
-              title: '',
-              dataIndex: 'attribute',
-              key: 'attribute',
-              align: 'left',
-              width: isMobile ? '0px' : '30%',
-              className: 'icon-padding white',
-              render: (value: string) => {
-                return `${mapKeyToAttribute(value)}:`;
+        </div>
+        <div className="flex-row">
+          <Table
+            showHeader={false}
+            style={{ width: '100%' }}
+            onRow={(row: ManageAntRow) => ({
+              className: row.attribute === editingField ? 'active-row' : '',
+            })}
+            scroll={{ x: true }}
+            columns={[
+              {
+                title: '',
+                dataIndex: 'attribute',
+                key: 'attribute',
+                align: 'left',
+                width: isMobile ? '0px' : '30%',
+                className: 'icon-padding white',
+                render: (value: string) => {
+                  return `${mapKeyToAttribute(value)}:`;
+                },
               },
-            },
-            {
-              title: '',
-              dataIndex: 'value',
-              key: 'value',
-              align: 'left',
-              width: '80%',
-              className: 'white',
-              render: (value: string | number, row: any) => {
-                if (row.attribute === 'status')
-                  return (
-                    <>
-                      {/* TODO: add label for mobile view */}
-                      <TransactionStatus confirmations={+value} />
-                    </>
-                  );
-                if (row.editable)
-                  return (
-                    <>
-                      {/* TODO: add label for mobile view */}
+              {
+                title: '',
+                dataIndex: 'value',
+                key: 'value',
+                align: 'left',
+                width: '80%',
+                className: 'white',
+                render: (value: string | number, row: any) => {
+                  if (row.attribute === 'status')
+                    return (
+                      <>
+                        {/* TODO: add label for mobile view */}
+                        <TransactionStatus confirmations={+value} />
+                      </>
+                    );
+                  if (row.editable)
+                    return (
+                      <>
+                        {/* TODO: add label for mobile view */}
 
-                      <ValidationInput
-                        showValidationIcon={true}
-                        showValidationOutline={true}
-                        inputId={row.attribute + '-input'}
-                        inputType={
-                          row.attribute === 'ttlSeconds' ? 'number' : undefined
-                        }
-                        minNumber={100}
-                        maxNumber={1000000}
-                        onClick={() => {
-                          setEditingField(row.attribute);
-                          setModifiedValue(value);
-                        }}
-                        inputClassName={'flex'}
-                        inputCustomStyle={{
-                          width: '100%',
-                          border: 'none',
-                          overflow: 'hidden',
-                          fontSize: '16px',
-                          outline: 'none',
-                          borderRadius: 'var(--corner-radius)',
-                          background:
-                            editingField === row.attribute
-                              ? 'white'
-                              : 'transparent',
-                          color:
-                            editingField === row.attribute ? 'black' : 'white',
-                          padding:
-                            editingField === row.attribute
-                              ? '10px 40px 10px 10px'
-                              : '10px 0px',
-                          display: 'flex',
-                        }}
-                        disabled={editingField !== row.attribute}
-                        placeholder={`Enter a ${mapKeyToAttribute(
-                          row.attribute,
-                        )}`}
-                        value={
-                          editingField === row.attribute
-                            ? modifiedValue
-                            : row.value
-                        }
-                        setValue={(e) => {
-                          if (row.attribute === editingField) {
-                            setModifiedValue(e);
+                        <ValidationInput
+                          showValidationIcon={true}
+                          showValidationOutline={true}
+                          inputId={row.attribute + '-input'}
+                          inputType={
+                            row.attribute === 'ttlSeconds'
+                              ? 'number'
+                              : undefined
                           }
-                        }}
-                        validityCallback={(valid: boolean) => {
-                          row.isValid = valid;
-                        }}
-                        validationPredicates={
-                          modifiedValue &&
-                          (row.attribute === 'owner' ||
-                            row.attribute === 'controller' ||
-                            row.attribute === 'targetID')
-                            ? {
-                                [VALIDATION_INPUT_TYPES.ARWEAVE_ID]: (
-                                  id: string,
-                                ) => arweaveDataProvider.validateArweaveId(id),
-                              }
-                            : {}
-                        }
-                        maxLength={43}
-                      />
-                    </>
-                  );
-                return value;
-              },
-            },
-            {
-              title: '',
-              dataIndex: 'action',
-              key: 'action',
-              width: '10%',
-              align: 'right',
-              className: 'white',
-              render: (value: any, row: any) => {
-                //TODO: if it's got an action attached, show it
-                if (row.editable) {
-                  return (
-                    <>
-                      {editingField !== row.attribute ? (
-                        <button
+                          minNumber={100}
+                          maxNumber={1000000}
                           onClick={() => {
                             setEditingField(row.attribute);
-                            setModifiedValue(row.value);
+                            setModifiedValue(value);
                           }}
-                        >
-                          <PencilIcon
-                            style={{ width: '24', height: '24', fill: 'white' }}
-                          />
-                        </button>
-                      ) : (
-                        <button
-                          className="assets-manage-button"
-                          style={{
-                            backgroundColor: 'var(--accent)',
-                            borderColor: 'var(--accent)',
+                          inputClassName={'flex'}
+                          inputCustomStyle={{
+                            width: '100%',
+                            border: 'none',
+                            overflow: 'hidden',
+                            fontSize: '16px',
+                            outline: 'none',
+                            borderRadius: 'var(--corner-radius)',
+                            background:
+                              editingField === row.attribute
+                                ? 'white'
+                                : 'transparent',
+                            color:
+                              editingField === row.attribute
+                                ? 'black'
+                                : 'white',
+                            padding:
+                              editingField === row.attribute
+                                ? '10px 40px 10px 10px'
+                                : '10px 0px',
+                            display: 'flex',
                           }}
-                          onClick={() => {
-                            const payload =
-                              getTransactionPayloadByInteractionType(
-                                CONTRACT_TYPES.ANT,
-                                row.interactionType,
-                                modifiedValue?.toString(),
-                              );
-                            if (payload && row.interactionType && id) {
-                              // eslint-disable-next-line
-                              const { assetId, functionName, ...data } =
-                                payload;
-                              dispatchTransactionState({
-                                type: 'setContractType',
-                                payload: CONTRACT_TYPES.ANT,
-                              });
-                              dispatchTransactionState({
-                                type: 'setInteractionType',
-                                payload: row.interactionType,
-                              });
-                              dispatchTransactionState({
-                                type: 'setTransactionData',
-                                payload: {
-                                  assetId: id,
-                                  functionName,
-                                  ...data,
-                                },
-                              });
-                              navigate(`/transaction`, {
-                                state: `/manage/ants/${id}`,
-                              });
+                          disabled={editingField !== row.attribute}
+                          placeholder={`Enter a ${mapKeyToAttribute(
+                            row.attribute,
+                          )}`}
+                          value={
+                            editingField === row.attribute
+                              ? modifiedValue
+                              : row.value
+                          }
+                          setValue={(e) => {
+                            if (row.attribute === editingField) {
+                              setModifiedValue(e);
                             }
                           }}
-                        >
-                          Save
-                        </button>
-                      )}
-                    </>
-                  );
-                }
-                if (row.action) {
-                  return (
-                    <button
-                      onClick={row.action.fn}
-                      className="assets-manage-button"
-                    >
-                      {row.action.title}
-                    </button>
-                  );
-                }
-                return value;
+                          validityCallback={(valid: boolean) => {
+                            row.isValid = valid;
+                          }}
+                          validationPredicates={
+                            modifiedValue &&
+                            (row.attribute === 'owner' ||
+                              row.attribute === 'controller' ||
+                              row.attribute === 'targetID')
+                              ? {
+                                  [VALIDATION_INPUT_TYPES.ARWEAVE_ID]: (
+                                    id: string,
+                                  ) =>
+                                    arweaveDataProvider.validateArweaveId(id),
+                                }
+                              : {}
+                          }
+                          maxLength={43}
+                        />
+                      </>
+                    );
+                  return value;
+                },
               },
-            },
-          ]}
-          data={rows}
-        />
+              {
+                title: '',
+                dataIndex: 'action',
+                key: 'action',
+                width: '10%',
+                align: 'right',
+                className: 'white',
+                render: (value: any, row: any) => {
+                  //TODO: if it's got an action attached, show it
+                  if (row.editable) {
+                    return (
+                      <>
+                        {editingField !== row.attribute ? (
+                          <button
+                            onClick={() => {
+                              setEditingField(row.attribute);
+                              setModifiedValue(row.value);
+                            }}
+                          >
+                            <PencilIcon
+                              style={{
+                                width: '24',
+                                height: '24',
+                                fill: 'white',
+                              }}
+                            />
+                          </button>
+                        ) : (
+                          <button
+                            className="assets-manage-button"
+                            style={{
+                              backgroundColor: 'var(--accent)',
+                              borderColor: 'var(--accent)',
+                            }}
+                            onClick={() => {
+                              const payload =
+                                getTransactionPayloadByInteractionType(
+                                  CONTRACT_TYPES.ANT,
+                                  row.interactionType,
+                                  modifiedValue?.toString(),
+                                );
+                              if (payload && row.interactionType && id) {
+                                // eslint-disable-next-line
+                                const { assetId, functionName, ...data } =
+                                  payload;
+                                dispatchTransactionState({
+                                  type: 'setContractType',
+                                  payload: CONTRACT_TYPES.ANT,
+                                });
+                                dispatchTransactionState({
+                                  type: 'setInteractionType',
+                                  payload: row.interactionType,
+                                });
+                                dispatchTransactionState({
+                                  type: 'setTransactionData',
+                                  payload: {
+                                    assetId: id,
+                                    functionName,
+                                    ...data,
+                                  },
+                                });
+                                navigate(`/transaction`, {
+                                  state: `/manage/ants/${id}`,
+                                });
+                              }
+                            }}
+                          >
+                            Save
+                          </button>
+                        )}
+                      </>
+                    );
+                  }
+                  if (row.action) {
+                    return (
+                      <button
+                        onClick={row.action.fn}
+                        className="assets-manage-button"
+                      >
+                        {row.action.title}
+                      </button>
+                    );
+                  }
+                  return value;
+                },
+              },
+            ]}
+            data={rows}
+          />
+        </div>
       </div>
-    </div>
+      {showTransferAntModal ? (
+        <TransferAntModal
+          showModal={() => setShowTransferAntModal(false)}
+          antId={new ArweaveTransactionID(id!)}
+        />
+      ) : (
+        <></>
+      )}
+    </>
   );
 }
