+++ src/components/modals/index.ts
@@ -1,6 +1,6 @@
 import ConnectWalletModal from './ConnectWalletModal/ConnectWalletModal';
 import CreateAntModal from './CreateAntModal/CreateAntModal';
 import ManageAntModal from './ManageAntModal/ManageAntModal';
-import TransactionModal from './TransactionModal/TransactionModal';
+import TransferAntModal from './TransferAntModal/TransferAntModal';
 
-export { CreateAntModal, ConnectWalletModal, ManageAntModal, TransactionModal };
+export { CreateAntModal, ConnectWalletModal, ManageAntModal, TransferAntModal };
